pyproject.toml
[project]
name = "eldas-backend"
version = "0.1.0"
description = "Production-ready FastAPI backend for Eldas."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.14.0",
    "boto3>=1.35.0",
    "email-validator>=2.2.0",
    "fastapi>=0.115.0",
    "httpx>=0.28.0",
    "openai>=1.58.0",
    "passlib[bcrypt]>=1.7.4",
    "pillow>=11.0.0",
    "psycopg[binary]>=3.2.0",
    "pydantic-settings>=2.7.0",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.12",
    "pytesseract>=0.3.13",
    "sqlalchemy>=2.0.36",
    "uvicorn[standard]>=0.32.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
]

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*"]
